+++ src/Integration/Bricks/Main.php
@@ -32,7 +32,9 @@ class Main implements IntegrationInterface
             add_filter('f!windpress/core/runtime:append_header.ubiquitous_panel.is_prevent_load', fn (bool $is_prevent_load): bool => $this->is_prevent_load($is_prevent_load));
             add_filter('f!windpress/core/runtime:append_header.exclude_admin', fn (bool $is_exclude_admin): bool => $this->is_exclude_admin($is_exclude_admin));
 
-            new Editor();
+            if (Config::get(sprintf('integration.%s.editor.enabled', $this->get_name()), true)) {
+                new Editor();
+            }
         }
     }
 
@@ -58,7 +60,10 @@ class Main implements IntegrationInterface
             'id' => $this->get_name(),
             'name' => __('Bricks Builder', 'windpress'),
             'description' => __('Bricks Builder integration', 'windpress'),
-            'callback' => Compile::class,
+            'callback' => Config::get(sprintf('integration.%s.compile.enabled', $this->get_name()), true)
+                    ? Compile::class
+                    : static fn() => []
+            ,
             'enabled' => $this->is_enabled(),
             'type' => 'theme',
             'homepage' => 'https://bricksbuilder.io/?ref=windpress',