+++ src/Integration/Builderius/Main.php
@@ -30,7 +30,10 @@ class Main implements IntegrationInterface
             add_filter('f!windpress/core/runtime:is_prevent_load', fn (bool $is_prevent_load): bool => $this->is_prevent_load($is_prevent_load));
             add_filter('f!windpress/core/runtime:append_header.ubiquitous_panel.is_prevent_load', fn (bool $is_prevent_load): bool => $this->is_prevent_load($is_prevent_load));
             add_filter('f!windpress/core/runtime:append_header.exclude_admin', fn (bool $is_exclude_admin): bool => $this->is_exclude_admin($is_exclude_admin));
-            new Editor();
+            
+            if (Config::get(sprintf('integration.%s.editor.enabled', $this->get_name()), true)) {
+                new Editor();
+            }
         }
     }
 
@@ -56,7 +59,10 @@ class Main implements IntegrationInterface
             'id' => $this->get_name(),
             'name' => __('Builderius', 'windpress'),
             'description' => __('Builderius integration', 'windpress'),
-            'callback' => Compile::class,
+            'callback' => Config::get(sprintf('integration.%s.compile.enabled', $this->get_name()), true)
+                    ? Compile::class
+                    : static fn() => []
+            ,
             'enabled' => $this->is_enabled(),
             'type' => 'plugin',
             'homepage' => 'https://builderius.io/?referral=afdfca82c8',