+++ src/Integration/Elementor/Main.php
@@ -53,7 +53,10 @@ class Main implements IntegrationInterface
             'id' => $this->get_name(),
             'name' => __('Elementor', 'windpress'),
             'description' => __('Elementor integration', 'windpress'),
-            'callback' => Compile::class,
+            'callback' => Config::get(sprintf('integration.%s.compile.enabled', $this->get_name()), true)
+                    ? Compile::class
+                    : static fn() => []
+            ,
             'enabled' => $this->is_enabled(),
             'type' => 'plugin',
             'homepage' => 'https://be.elementor.com/visit/?bta=209150&brand=elementor',